KP-14263 Optimize deletion of a questionnaire

+++ src/Core/BoundedContexts/Headquarters/WB.Core.BoundedContexts.Headquarters/Questionnaires/Jobs/DeleteQuestionnaireJob.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using Quartz;
 using WB.Core.BoundedContexts.Headquarters.Services.DeleteQuestionnaireTemplate;
@@ -27,6 +28,9 @@ namespace WB.Core.BoundedContexts.Headquarters.Questionnaires.Jobs
 
             if(disabledNotDeletedQuestionnaire == null) return Task.CompletedTask;
 
+            if (!disabledNotDeletedQuestionnaire.DisabledBy.HasValue)
+                throw new ArgumentException();
+                
             deleteQuestionnaireService.DeleteInterviewsAndQuestionnaireAfter(disabledNotDeletedQuestionnaire.QuestionnaireId, disabledNotDeletedQuestionnaire.Version, disabledNotDeletedQuestionnaire.DisabledBy);
 
             return Task.CompletedTask;
